feat: add create session dialog and refactor session handling in chat components
