+++ web/src/features/chat/chat.tsx
@@ -57,9 +57,7 @@ type ChatWorkspaceProps = {
   /** Whether waiting for the first response after a prompt is sent */
   isAwaitingFirstResponse?: boolean;
   /** Create a new session when none is selected */
-  onCreateSession?: () => Promise<void>;
-  /** Whether a new session is being created */
-  isCreatingSession?: boolean;
+  onCreateSession?: () => void;
 };
 
 type ToolApproval = NonNullable<LiveMessage["toolCall"]>["approval"];
@@ -83,7 +81,6 @@ export const ChatWorkspace = memo(function ChatWorkspaceComponent({
   isUploadingFiles = false,
   isAwaitingFirstResponse = false,
   onCreateSession,
-  isCreatingSession = false,
 }: ChatWorkspaceProps): ReactElement {
   const [blocksExpanded, setBlocksExpanded] = useState(false);
   const [pendingApprovalMap, setPendingApprovalMap] = useState<
@@ -159,7 +156,6 @@ export const ChatWorkspace = memo(function ChatWorkspaceComponent({
             canRespondToApproval={Boolean(onApprovalResponse)}
             blocksExpanded={blocksExpanded}
             onCreateSession={onCreateSession}
-            isCreatingSession={isCreatingSession}
           />
         </div>
